toy_project/requirements.txt
pandas
celluloid
numpy
matplotlib
python-dateutil
